+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## firestore-17.0.2.0, database-kotlin-16.0.1.2 *(2018/07/02)*
+
+Built and tested with:
+- Firebase Android SDK
+  - `firebase-firestore:17.0.1`
+  - `firebase-database:16.0.1`
+- Kotlin 1.2.41
+- RxJava2 2.1.8
+
+New module:
+- `firebase-firestore` and `firebase-firestore-kotlin` for [Cloud Firestore](https://firebase.google.com/docs/firestore/) support
+
+Fixed:
+- Misspells on `childEvents()` extension method (in `firebase-database-kotlin` module)
+
 ## core-16.0.1.0, auth-16.0.2.0 *(2018/06/21)*
 
 Built and tested with: